+++ srcs_bonus/config_parse/parse_map.c
@@ -29,7 +29,7 @@ void	add_line_to_map(char *line, t_cub *cub)
 	(cub->map->field)[i + 1] = NULL;
 }
 
-int		check_map(t_cub *cub)
+size_t check_map(t_cub *cub)
 {
 	if (!is_map_chars_correct(" 012NWES", cub))
 		return (0);
